Move *.c and *.h to the code folder and update Makefile to include tests
