+++ code/vehicle.h
@@ -10,4 +10,14 @@ typedef struct {
   bool requestCarStart;
 } Vehicle;
 
+typedef enum {
+  DRIVE_INTENTION_NOTHING = 0,
+  DRIVE_INTENTION_REDUCE_SPEED = 1,
+  DRIVE_INTENTION_KEEP_SPEED = 2,
+  DRIVE_INTENTION_INCREASE_SPEED = 3,
+ 
+} DriverIntention;
+
+DriverIntention getDriverIntention(Vehicle vehicle);
+
 #endif // VEHICLE_H
